Made some benchamarking.
It seams that Jax is faster however, I think that this is an arthefact.
If I increase the K dimension then the SDFG versions stays roughly the same, but JAX increases.
I think that that what we see is the setupcode, however, I have to verify this by makeing a scaling analysis.