Now the normal tests works for GPU.
However something does not work, for example:
- Scalars as arguments.
The will be automatically turned into an array with shape `(1,)` and a warning will be issued.
- Constants are not supported.
I have no idea why but they do not work, they are simply forbidden.
But it seems that DaCe does not implement them as well, I have no idea why.