Fixed some variable nameing issue.
JAX might generate variable names such as `if` or `do` or ewben `register` which are C++ keywords.
This commit will rename such variables to prevent a conflict.

The fix is overcomplicated and I do not like it at all.
I have to find a bit better solution, whcih is cleaner.