This fixes the strange error caused with the auto optimization and GPU.
One of the main issue that the optimizations created some stuff, probably temporaries, that were not GPU.
An example are maps that were created that had a scheduling of CPU and GPU, another problem was that while `c` is on GPU but the memlet `c[0, 0, 0]` was (somehow) seen as to be on CPU.
These caused the validations after the GPU transformation or the auto optimization to fail.
The trick now is that the GPU transformation is done at last, _and_ all validations are disable.
To ensure that everything is working aft the end final validation is performed.

However, I do not like this solution, although it makes quite sense.