Fixed the ICON benchmark.
However, there is an error in the `ITIR -> SDFG` version, the same that is Edoardo (https://github.com/spcl/dace/pull/1480) is working on.
Soi it is deactivated for the moment.
But, it was not present on CPU, or was ist.

But I am not sure if it still works on CPU?