Benchmarked the GPU implementation.
In the CPU JAX shows a clear inscrese in runtime but not in the GPU.
I think that this is related to the compute grids on the GPU.
Since we just use the default JAX is most likely adapting them.
Thus as we increase the work JAX also makes more resources avaiable.
At least this is my tentative explanation.