Redesigned the Jax to SDFG translator.
The new design is an important step towards supporting lambda.
The translator that was written before is now split into two parts.
The base part `JaxprBaseTranslator` handles the bulk of the translation, however, it does not make optimization or creates the return value.
The most important feature is that it provides a `clone()` function which gives a new translator that can then be used to translate lambdas.

The old translator `JaxprToSDFG` is still there and is derived from the `JaxprBaseTranslator`, which was done only for compatibility, I should have probably made it an internal object.
This class then performs the interfacing stuff, such as generating the return statements and perform the optimization.
However, while the base is not that bad this part of the code is quite bad and I should probably rewrite it.